# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CyberCat",
    "description": "Threat-informed automated incident response platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/incidents": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_v1_incidents_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "severity_gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity Gte"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "opened_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Opened After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/{incident_id}/transitions": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Transition Incident",
        "operationId": "transition_incident_v1_incidents__incident_id__transitions_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransitionOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/v1/incidents/{incident_id}/notes": {
      "post": {
        "tags": [
          "incidents"
        ],
        "summary": "Add Note",
        "operationId": "add_note_v1_incidents__incident_id__notes_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteRef"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntityKind": {
        "type": "string",
        "enum": [
          "user",
          "host",
          "ip",
          "process",
          "file",
          "observable"
        ],
        "title": "EntityKind"
      },
      "EntityRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/EntityKind"
          },
          "natural_key": {
            "type": "string",
            "title": "Natural Key"
          },
          "attrs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attrs"
          },
          "role_in_incident": {
            "type": "string",
            "title": "Role In Incident"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "natural_key",
          "attrs",
          "role_in_incident"
        ],
        "title": "EntityRef"
      },
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "confidence": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "correlator_rule": {
            "type": "string",
            "title": "Correlator Rule"
          },
          "correlator_version": {
            "type": "string",
            "title": "Correlator Version"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityRef"
            },
            "type": "array",
            "title": "Entities"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/DetectionRef"
            },
            "type": "array",
            "title": "Detections"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            },
            "type": "array",
            "title": "Timeline"
          },
          "attack": {
            "items": {
              "$ref": "#/components/schemas/AttackRef"
            },
            "type": "array",
            "title": "Attack"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionSummary"
            },
            "type": "array",
            "title": "Actions"
          },
          "transitions": {
            "items": {
              "$ref": "#/components/schemas/TransitionRef"
            },
            "type": "array",
            "title": "Transitions"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteRef"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "kind",
          "status",
          "severity",
          "confidence",
          "rationale",
          "opened_at",
          "updated_at",
          "closed_at",
          "correlator_rule",
          "correlator_version",
          "entities",
          "detections",
          "timeline",
          "attack",
          "actions",
          "transitions",
          "notes"
        ],
        "title": "IncidentDetail"
      },
      "IncidentKind": {
        "type": "string",
        "enum": [
          "identity_compromise",
          "endpoint_compromise",
          "identity_endpoint_chain",
          "unknown"
        ],
        "title": "IncidentKind"
      },
      "IncidentList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "IncidentList"
      },
      "IncidentStatus": {
        "type": "string",
        "enum": [
          "new",
          "triaged",
          "investigating",
          "contained",
          "resolved",
          "closed",
          "reopened"
        ],
        "title": "IncidentStatus"
      },
      "IncidentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "$ref": "#/components/schemas/IncidentKind"
          },
          "status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "confidence": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Confidence"
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "title": "Opened At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "entity_count": {
            "type": "integer",
            "title": "Entity Count"
          },
          "detection_count": {
            "type": "integer",
            "title": "Detection Count"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "primary_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary User"
          },
          "primary_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Host"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "kind",
          "status",
          "severity",
          "confidence",
          "opened_at",
          "updated_at",
          "entity_count",
          "detection_count",
          "event_count",
          "primary_user",
          "primary_host"
        ],
        "title": "IncidentSummary"
      },
...
      "NoteIn": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "NoteIn"
      },
      "NoteRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "body",
          "author",
          "created_at"
        ],
        "title": "NoteRef"
      },
...
      "Severity": {
        "type": "string",
        "enum": [
          "info",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity"
      },
...
      "TransitionIn": {
        "properties": {
          "to_status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "to_status"
        ],
        "title": "TransitionIn"
      },
      "TransitionOut": {
        "properties": {
          "incident_id": {
            "type": "string",
            "format": "uuid",
            "title": "Incident Id"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IncidentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/IncidentStatus"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          }
        },
        "type": "object",
        "required": [
          "incident_id",
          "from_status",
          "to_status",
          "at"
        ],
        "title": "TransitionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}